+++ CHANGELOG.md
@@ -1,7 +1,24 @@
 # Changelog
+## 2.0.7
+- AtsReception changes:
+    - Added [AtsEntry] linked to the [AtsReception].
+- AtsReceptionInput changes:
+    - Added [AtsReception] operation time
+    - Added App used to create the [AtsReception]
+- AtsReceptionProductInput changes:
+    - Removed assetId, fuelType and volumeBought
+## 2.0.6
+- Added `dynamicIcon` of type `Avatar` to `InboundProtocol`
+
+## 2.0.5
+- Modified `Model` entity to support more fields.
+
+## 2.0.4
+- added speed to `WorkspaceTripPoint` entity
 
 ## 2.0.3
 - TimestampConverter and TimestampConverterOrNull now can recive any type of num 
+
 ## 2.0.2
 - Some linting issues fixed
 
@@ -10,4 +27,3 @@
 
 ## 2.0.0
 - Initial public release
-